pyproject.toml
[project]
name = "synapse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "langchain==0.3.26",
    "langchain-community==0.3.26",
    "langchain-core==0.3.66",
    "langchain-ollama==0.3.3",
    "langgraph==0.5.0",
    "langchain-together==0.3.0",
    "langgraph-supervisor==0.0.27",
    "pybreaker>=1.4.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "uvicorn>=0.38.0",
]
